Clean up workspace header and move entity diagram
